+++ src/domain/health-check/app-health.interface.ts
@@ -1,0 +1,7 @@
+import { Status } from '../../application/enums/status'
+
+export interface AppHealth {
+  uptime: number
+  status: Status
+  timestamp: number
+}